+++ src/ntfsclone.h
@@ -71,5 +71,5 @@ typedef struct ntfsclone_atom {
 	unsigned char	ncau_used_cluster[sizeof(u_int64_t)];
     }		nca_union;
 } __attribute__((__packed__)) ntfsclone_atom_t;
-#define	ATOM_TO_DATA_OFFSET	((u_int32_t) &(((ntfsclone_atom_t *) NULL)->nca_union.ncau_used_cluster[0]))
+#define	ATOM_TO_DATA_OFFSET	((u_int64_t) &(((ntfsclone_atom_t *) NULL)->nca_union.ncau_used_cluster[0]))
 #endif	/* _NTFSCLONE_H_ */
